feature: build script and the generated PDF
Having to install all these packages when probably they will only be
used once is not worth it. Also, depending on the operating system's
version, the builder will have to find alternative packages or fix the
build itself.

The build script should help anyone who wants to build the project to
build it in their machines, if they so want to.

However, even to avoid having to do that, I think it is a good idea to
upload the generated PDF itself too for convenience.